+++ src/Gnugat/Redaktilo/Replace/ReplaceEngine.php
@@ -9,8 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Gnugat\Redaktilo\Engine;
+namespace Gnugat\Redaktilo\Replace;
 
+use Gnugat\Redaktilo\Engine\NotSupportedException;
 use Gnugat\Redaktilo\Replace\ReplaceStrategy;
 
 /**
@@ -21,11 +22,11 @@ use Gnugat\Redaktilo\Replace\ReplaceStrategy;
  */
 class ReplaceEngine
 {
-    /** @var array of ReplaceStrategy */
+    /** @var ReplaceStrategy[] */
     private $replaceStrategies = array();
 
     /**
-     * @param ReplaceStrategy $searchEngine
+     * @param ReplaceStrategy $replaceStrategy
      *
      * @api
      */